Remove 'Please select a category' error that appears incorrectly. (#3440)
* Fix category selection for default fee in edit view.
* Prevent listings been deleted in edit view when “Clear Form” button is clicked.

+++ includes/views/submit_listing.php
@@ -75,8 +75,12 @@ class WPBDP__Views__Submit_Listing extends WPBDP__Authenticated_Listing_View {
 
         // Handle "Clear Form" request.
         if ( ! empty( $_POST ) && ! empty( $_POST['reset'] ) && 'reset' == $_POST['reset'] ) {
-            wp_delete_post( $this->listing->get_id(), true );
-            return $this->_redirect( wpbdp_url( 'submit_listing' ) );
+            if ( ! $this->editing ) {
+                wp_delete_post( $this->listing->get_id(), true );
+                return $this->_redirect( wpbdp_url( 'submit_listing' ) );
+            }
+
+            return $this->_redirect( wpbdp_url( 'edit_listing', $this->listing->get_id() ) );
         }
 
         if ( ! $this->editing && 'auto-draft' != get_post_status( $this->listing->get_id() ) ) {
